# api/01-openapi-spec.yaml
openapi: 3.0.3
info:
  title: Ride Circle API
  description: |
    Comprehensive API for the Ride Circle peer-to-peer car rental platform.
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    
    ## Rate Limiting
    - Public endpoints: 100 requests/minute
    - Authenticated endpoints: 1000 requests/minute
    - Admin endpoints: 5000 requests/minute
    
    ## Multi-tenancy
    All requests must include the tenant context via subdomain (mumbai.ridecircle.in) 
    or X-Tenant-Code header.
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format with additional context.
    
  version: 1.0.0
  contact:
    name: Ride Circle API Support
    email: api-support@ridecircle.in
  license:
    name: Proprietary
servers:
  - url: https://api.ridecircle.in/v1
    description: Production server
  - url: https://staging-api.ridecircle.in/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /vehicles:
    get:
      tags: [Vehicles]
      summary: Search vehicles
      description: Search and filter available vehicles
      security: []
      parameters:
        - name: location
          in: query
          description: Location coordinates (lat,lng)
          schema:
            type: string
            pattern: '^-?\d+\.?\d*,-?\d+\.?\d*$'
            example: "19.0760,72.8777"
        - name: radius
          in: query
          description: Search radius in kilometers
          schema:
            type: number
            minimum: 1
            maximum: 50
            default: 10
        - name: startDate
          in: query
          description: Booking start date
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Booking end date
          schema:
            type: string
            format: date-time
        - name: categoryId
          in: query
          description: Vehicle category filter
          schema:
            type: string
            format: uuid
        - name: makeId
          in: query
          description: Vehicle make filter
          schema:
            type: string
            format: uuid
        - name: priceMin
          in: query
          description: Minimum price per day
          schema:
            type: number
            minimum: 0
        - name: priceMax
          in: query
          description: Maximum price per day
          schema:
            type: number
        - name: transmission
          in: query
          description: Transmission type
          schema:
            type: string
            enum: [manual, automatic, cvt]
        - name: fuelType
          in: query
          description: Fuel type
          schema:
            type: string
            enum: [petrol, diesel, cng, electric, hybrid]
        - name: features
          in: query
          description: Required features (comma-separated)
          schema:
            type: string
            example: "bluetooth,ac,gps"
        - name: sortBy
          in: query
          description: Sort criteria
          schema:
            type: string
            enum: [price, rating, distance, popularity]
            default: distance
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Vehicles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      tags: [Vehicles]
      summary: Add new vehicle
      description: Add a new vehicle to the platform (Host only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreate'
      responses:
        '201':
          description: Vehicle created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        errors:
          type: array
          items:
            type: string
        code:
          type: string
          example: "VALIDATION_ERROR"
        timestamp:
          type: string
          format: date-time
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false

    # Authentication schemas
...
    Vehicle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hostId:
          type: string
          format: uuid
        make:
          type: string
          example: "Honda"
        model:
          type: string
          example: "City"
        year:
          type: integer
          example: 2020
        registrationNumber:
          type: string
          example: "MH01AB1234"
        color:
          type: string
          example: "Silver"
        fuelType:
          type: string
          enum: [petrol, diesel, cng, electric, hybrid]
        transmission:
          type: string
          enum: [manual, automatic, cvt]
        seatingCapacity:
          type: integer
          example: 5
        status:
          type: string
          enum: [draft, pending_approval, active, inactive, maintenance, retired]
        location:
          type: object
          properties:
            lat:
              type: number
              example: 19.0760
            lng:
              type: number
              example: 72.8777
        address:
          type: object
          properties:
            street:
              type: string
            area:
              type: string
            city:
              type: string
            state:
              type: string
            pincode:
              type: string
        basePricePerDay:
          type: number
          example: 2500
        basePricePerHour:
          type: number
          example: 350
        securityDeposit:
          type: number
          example: 5000
        features:
          type: array
          items:
            type: string
          example: ["bluetooth", "ac", "gps"]
        averageRating:
          type: number
          example: 4.5
        totalTrips:
          type: integer
          example: 25
        createdAt:
          type: string
          format: date-time
    VehicleCreate:
      type: object
      required: [modelId, registrationNumber, year, fuelType, transmission, seatingCapacity, location, address, basePricePerDay, securityDeposit]
      properties:
        modelId:
          type: string
          format: uuid
        registrationNumber:
          type: string
        year:
          type: integer
          minimum: 2000
          maximum: 2025
        color:
          type: string
        fuelType:
          type: string
          enum: [petrol, diesel, cng, electric, hybrid]
        transmission:
          type: string
          enum: [manual, automatic, cvt]
        seatingCapacity:
          type: integer
          minimum: 2
          maximum: 8
        location:
          type: object
          required: [lat, lng]
          properties:
            lat:
              type: number
            lng:
              type: number
        address:
          type: object
          required: [street, city, state, pincode]
          properties:
            street:
              type: string
            area:
              type: string
            city:
              type: string
            state:
              type: string
            pincode:
              type: string
        basePricePerDay:
          type: number
          minimum: 500
        basePricePerHour:
          type: number
          minimum: 50
        securityDeposit:
          type: number
          minimum: 1000
        features:
          type: array
          items:
            type: string
        rules:
          type: object
          properties:
            smoking:
              type: boolean
            pets:
              type: boolean
            mileageLimit:
              type: integer
...
    VehicleSearchResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
        filters:
          type: object
          properties:
            appliedFilters:
              type: object
            availableFilters:
              type: object
...
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Validation failed"
            errors: ["Email is required", "Phone number is invalid"]
            code: "VALIDATION_ERROR"
            timestamp: "2024-01-15T10:30:00Z"
...
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Access denied"
            code: "FORBIDDEN"
            timestamp: "2024-01-15T10:30:00Z"
...